+++ repoze/what/plugins/quickstart/config.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ##############################################################################
 #
-# Copyright (c) 2009, Gustavo Narea <me@gustavonarea.net>.
+# Copyright (c) 2009-2010, Gustavo Narea <me@gustavonarea.net>.
 # All Rights Reserved.
 #
 # This software is subject to the provisions of the BSD-like license at
@@ -132,6 +132,7 @@ class _AuthConf(object):
         """
         # Adding general options:
         self._add_object("general", "dbsession")
+        self._add_string("general", "charset")
         
         # Adding authentication options:
         self._add_object("authentication", "user_class")